ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
langchain==0.1.0
openai==1.6.1
python-dotenv==1.0.0
chromadb==0.4.18
tiktoken==0.5.2
pypdf==3.17.4
python-multipart==0.0.6
